API/requirements.txt
Flask
flask-restx
psycopg2-binary
SQLAlchemy
requests